+++ spec/spec/hyperlink.js
@@ -46,14 +46,14 @@ describe('Hyperlink', function() {
     describe('linkTo()', function() {
       it('wraps the called element in a link with given url', function() {
         element.linkTo(url)
-        expect(element.parent.attr('href')).toBe(url)
+        expect(element.parent().attr('href')).toBe(url)
       })
       it('wraps the called element in a link with given block', function() {
         element.linkTo(function(link) {
           link.to(url).target('_blank')
         })
-        expect(element.parent.attr('href')).toBe(url)
-        expect(element.parent.attr('target')).toBe('_blank')
+        expect(element.parent().attr('href')).toBe(url)
+        expect(element.parent().attr('target')).toBe('_blank')
       })
     })
   })