+++ src/selector.js
@@ -1,5 +1,20 @@
 // Method for getting an element by id
 SVG.get = function(id) {
   var node = document.getElementById(idFromReference(id) || id)
-  if (node) return node.instance
+  if (node) return SVG.adopt(node)
 }
+
+// Select elements by query string
+SVG.select = function(query, parent) {
+  return SVG.utils.map((parent || document).querySelectorAll(query), function(node) {
+    return SVG.adopt(node)
+  })
+}
+
+SVG.extend(SVG.Parent, {
+  // Scoped select method
+  select: function(query) {
+    return SVG.select(query, this.node)
+  }
+
+})