+++ src/utilities.js
@@ -1,0 +1,14 @@
+SVG.utils = {
+  // Map function
+  map: function(array, block) {
+    var i
+      , il = array.length
+      , result = []
+
+    for (i = 0; i < il; i++)
+      result.push(block(array[i]))
+    
+    return result
+  }
+
+}